refactor(@angular/cli): remove old `CommandModule` and `SchematicCommand` implementations
Use the newly implemented `CommandModule` and `SchematicsCommandModule`.
