+++ packages/angular/cli/src/commands/new/cli.ts
@@ -17,9 +17,10 @@ import {
   SchematicsCommandArgs,
   SchematicsCommandModule,
 } from '../../command-builder/schematics-command-module';
-import { NewCommand } from './new-impl';
+import { ensureCompatibleNpm } from '../../utilities/package-manager';
+import { VERSION } from '../../utilities/version';
 
-export interface NewCommandArgs extends SchematicsCommandArgs {
+interface NewCommandArgs extends SchematicsCommandArgs {
   collection?: string;
 }
 
@@ -27,8 +28,9 @@ export class NewCommandModule
   extends SchematicsCommandModule
   implements CommandModuleImplementation<NewCommandArgs>
 {
-  protected override schematicName = 'ng-new';
+  private readonly schematicName = 'ng-new';
   static override scope = CommandScope.Out;
+  protected override allowPrivateSchematics = true;
 
   command = 'new [name]';
   aliases = 'n';
@@ -36,18 +38,55 @@ export class NewCommandModule
   longDescriptionPath?: string | undefined;
 
   override async builder(argv: Argv): Promise<Argv<NewCommandArgs>> {
-    const baseYargs = await super.builder(argv);
-
-    return baseYargs.option('collection', {
+    const localYargs = (await super.builder(argv)).option('collection', {
       alias: 'c',
       describe: 'A collection of schematics to use in generating the initial application.',
       type: 'string',
     });
+
+    const {
+      options: { collectionNameFromArgs },
+    } = this.context.args;
+
+    const collectionName =
+      typeof collectionNameFromArgs === 'string'
+        ? collectionNameFromArgs
+        : await this.getDefaultSchematicCollection();
+
+    const workflow = await this.getOrCreateWorkflowForBuilder(collectionName);
+    const collection = workflow.engine.createCollection(collectionName);
+    const options = await this.getSchematicOptions(collection, this.schematicName, workflow);
+
+    return this.addSchemaOptionsToCommand(localYargs, options);
   }
 
-  run(options: Options<NewCommandArgs> & OtherOptions): number | void | Promise<number | void> {
-    const command = new NewCommand(this.context, 'new');
+  async run(options: Options<NewCommandArgs> & OtherOptions): Promise<number | void> {
+    // Register the version of the CLI in the registry.
+    const collectionName = options.collection ?? (await this.getDefaultSchematicCollection());
+    const workflow = await this.getOrCreateWorkflowForExecution(collectionName, options);
+    workflow.registry.addSmartDefaultProvider('ng-cli-version', () => VERSION.full);
 
-    return command.validateAndRun(options);
+    const { dryRun, force, interactive, defaults, collection, ...schematicOptions } = options;
+
+    // Compatibility check for NPM 7
+    if (
+      collectionName === '@schematics/angular' &&
+      !schematicOptions.skipInstall &&
+      (schematicOptions.packageManager === undefined || schematicOptions.packageManager === 'npm')
+    ) {
+      await ensureCompatibleNpm(this.context.root);
+    }
+
+    return this.runSchematic({
+      collectionName,
+      schematicName: this.schematicName,
+      schematicOptions,
+      executionOptions: {
+        dryRun,
+        force,
+        interactive,
+        defaults,
+      },
+    });
   }
 }
